+++ src/hooks/useLogin.js
@@ -1,0 +1,51 @@
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
+import { loginUser } from "../services/authService";
+
+const useLogin = () => {
+  const navigate = useNavigate();
+
+  const [formData, setFormData] = useState({
+    email: "",
+    password: "",
+  });
+
+  const [error, setError] = useState("");
+
+  const handleInputChange = (e) => {
+    const { name, value } = e.target;
+    setFormData((prev) => ({ ...prev, [name]: value }));
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setError("");
+
+    const { email, password } = formData;
+
+    if (!email || !password) {
+      setError("All fields are required");
+      return;
+    }
+
+    try {
+    //   const data = await loginUser(email, password);
+      await loginUser(email, password);
+      Swal.fire({
+        title: "Login Successful!",
+        text: "Welcome back!",
+        icon: "success",
+        confirmButtonText: "Continue",
+      }).then(() => {
+        navigate("/dashboard");
+      });
+    } catch (error) {
+      setError(error.message || "Something went wrong");
+    }
+  };
+
+  return { formData, error, handleInputChange, handleSubmit };
+};
+
+export default useLogin;
